requirements.txt
demoparser2
streamlit
matplotlib
numpy
pandas
seaborn
awpy2
plotly
playwright